fix memory corruption and feature matching accuracy
The distance between local descriptors should be float essentially,
but they had been casted to int.

There were also bugs in checking whether a match has been found,
which could cause invalid memory access.
(writing to CurrentFrame.mvpMapPoints[-1] in some cases)

+++ include/Matcher.h
@@ -14,13 +14,13 @@ namespace DXSLAM
 {
 
 class Matcher
-{    
+{
 public:
 
     Matcher(float nnratio=0.6, bool checkOri=true);
 
     // Computes the Hamming distance between two ORB descriptors
-    static int DescriptorDistance(const cv::Mat &a, const cv::Mat &b);
+    static float DescriptorDistance(const cv::Mat &a, const cv::Mat &b);
 
     // Search matches between Frame keypoints and projected MapPoints. Returns number of matches
     // Used to track the local map (Tracking)
@@ -60,8 +60,8 @@ public:
 
 public:
 
-    static const int TH_LOW;
-    static const int TH_HIGH;
+    static const float TH_LOW;
+    static const float TH_HIGH;
     static const int HISTO_LENGTH;
